+++ lib/cucumber/glue/proto_world.rb
@@ -1,7 +1,7 @@
 # frozen_string_literal: true
 
 require 'cucumber/gherkin/formatter/ansi_escapes'
-require 'cucumber/core/ast/data_table'
+require 'cucumber/core/test/data_table'
 
 module Cucumber
   module Glue
@@ -24,7 +24,7 @@ module Cucumber
       # @example Passing a multiline string
       #   step "the email should contain:", "Dear sir,\nYou've won a prize!\n"
       # @param [String] name The name of the step
-      # @param [String,Cucumber::Ast::DocString,Cucumber::Ast::Table] multiline_argument
+      # @param [String,Cucumber::Test::DocString,Cucumber::Ast::Table] multiline_argument
       def step(name, raw_multiline_arg = nil)
         super
       end
@@ -68,7 +68,7 @@ module Cucumber
       #   ])
       #
       def table(text_or_table, file = nil, line = 0)
-        location = !file ? Core::Ast::Location.of_caller : Core::Ast::Location.new(file, line)
+        location = !file ? Core::Test::Location.of_caller : Core::Test::Location.new(file, line)
         MultilineArgument::DataTable.from(text_or_table, location)
       end
 
@@ -137,12 +137,12 @@ module Cucumber
           end
 
           define_method(:step) do |name, raw_multiline_arg = nil|
-            location = Core::Ast::Location.of_caller
+            location = Core::Test::Location.of_caller
             runtime.invoke_dynamic_step(name, MultilineArgument.from(raw_multiline_arg, location))
           end
 
           define_method(:steps) do |steps_text|
-            location = Core::Ast::Location.of_caller
+            location = Core::Test::Location.of_caller
             runtime.invoke_dynamic_steps(steps_text, language, location)
           end