+++ lib/cucumber/glue/registry_and_more.rb
@@ -178,7 +178,7 @@ module Cucumber
     end
 
     def self.backtrace_line(proc, name)
-      location = Cucumber::Core::Ast::Location.from_source_location(*proc.source_location)
+      location = Cucumber::Core::Test::Location.from_source_location(*proc.source_location)
       "#{location}:in `#{name}'"
     end
   end