+++ spec/cucumber/formatter/json_spec.rb
@@ -121,7 +121,7 @@ module Cucumber
                       "line": 4,
                       "match": {"location": "spec/cucumber/formatter/json_spec.rb:100"},
                       "result": {"status": "failed",
-                                 "error_message": "no bananas (RuntimeError)\\n./spec/cucumber/formatter/json_spec.rb:100:in `/^there are bananas$/'\\nspec.feature:4:in `Given there are bananas'",
+                                 "error_message": "no bananas (RuntimeError)\\n./spec/cucumber/formatter/json_spec.rb:100:in `/^there are bananas$/'\\nspec.feature:4:in `there are bananas'",
                                  "duration": 1}}]}]}]})
           end
         end
@@ -159,7 +159,7 @@ module Cucumber
                       "line": 4,
                       "match": {"location": "spec/cucumber/formatter/json_spec.rb:138"},
                       "result": {"status": "pending",
-                                 "error_message": "TODO (Cucumber::Pending)\\n./spec/cucumber/formatter/json_spec.rb:138:in `/^there are bananas$/'\\nspec.feature:4:in `Given there are bananas'",
+                                 "error_message": "TODO (Cucumber::Pending)\\n./spec/cucumber/formatter/json_spec.rb:138:in `/^there are bananas$/'\\nspec.feature:4:in `there are bananas'",
                                  "duration": 1}}]}]}]})
           end
         end
@@ -309,7 +309,7 @@ module Cucumber
             Then(/^the monkey eats bananas$/) {}
           end
 
-          it 'the comments are included in the json data' do
+          it 'the comments are not included in the json data' do
             expect(load_normalised_json(@out)).to eq MultiJson.load(%(
               [{"id": "banana-party",
                 "uri": "spec.feature",
@@ -317,15 +317,11 @@ module Cucumber
                 "name": "Banana party",
                 "line": 2,
                 "description": "",
-                "comments": [{"value": "#feature comment",
-                              "line": 1}],
                 "elements":
                  [{"keyword": "Background",
                    "name": "There are bananas",
                    "line": 5,
                    "description": "",
-                   "comments": [{"value": "#background comment",
-                                 "line": 4}],
                    "type": "background",
                    "steps":
                     [{"keyword": "Given ",
@@ -339,15 +335,11 @@ module Cucumber
                    "name": "Monkey eats bananas",
                    "line": 9,
                    "description": "",
-                   "comments": [{"value": "#scenario comment",
-                                 "line": 8}],
                    "type": "scenario",
                    "steps":
                     [{"keyword": "Then ",
                       "name": "the monkey eats bananas",
                       "line": 11,
-                      "comments": [{"value": "#step comment1",
-                                    "line": 10}],
                       "match": {"location": "spec/cucumber/formatter/json_spec.rb:309"},
                       "result": {"status": "passed",
                                  "duration": 1}}]},
@@ -355,8 +347,6 @@ module Cucumber
                    "name": "There are bananas",
                    "line": 5,
                    "description": "",
-                   "comments": [{"value": "#background comment",
-                                 "line": 4}],
                    "type": "background",
                    "steps":
                     [{"keyword": "Given ",
@@ -370,19 +360,11 @@ module Cucumber
                    "name": "Monkey eats bananas",
                    "line": 22,
                    "description": "",
-                   "comments": [{"value": "#scenario outline comment",
-                                 "line": 13},
-                                {"value": "#examples table comment",
-                                 "line": 18},
-                                {"value": "#examples table row comment",
-                                 "line": 21}],
                    "type": "scenario",
                    "steps":
                     [{"keyword": "Then ",
                       "name": "the monkey eats bananas",
                       "line": 16,
-                      "comments": [{"value": "#step comment2",
-                                    "line": 15}],
                       "match": {"location": "spec/cucumber/formatter/json_spec.rb:309"},
                       "result": {"status": "passed",
                                  "duration": 1}}]}]}]))
@@ -426,7 +408,7 @@ module Cucumber
                       "doc_string": {"value": "the doc string",
                                      "content_type": "",
                                      "line": 5},
-                      "match": {"location": "spec/cucumber/formatter/json_spec.rb:404"},
+                      "match": {"location": "spec/cucumber/formatter/json_spec.rb:386"},
                       "result": {"status": "passed",
                                  "duration": 1}}]}]}]))
           end
@@ -464,7 +446,7 @@ module Cucumber
                       "name": "there are bananas",
                       "line": 4,
                       "output": ["from step"],
-                      "match": {"location": "spec/cucumber/formatter/json_spec.rb:444"},
+                      "match": {"location": "spec/cucumber/formatter/json_spec.rb:426"},
                       "result": {"status": "passed",
                                  "duration": 1}}]}]}]))
           end
@@ -575,7 +557,7 @@ module Cucumber
                       "line": 4,
                       "embeddings": [{"mime_type": "mime-type",
                                       "data": "YWJj"}],
-                      "match": {"location": "spec/cucumber/formatter/json_spec.rb:554"},
+                      "match": {"location": "spec/cucumber/formatter/json_spec.rb:536"},
                       "result": {"status": "passed",
                                  "duration": 1}}]}]}]))
           end
@@ -620,7 +602,7 @@ module Cucumber
                       "line": 4,
                       "embeddings": [{"mime_type": "image/png",
                                       "data": "Zm9v"}],
-                      "match": {"location": "spec/cucumber/formatter/json_spec.rb:593"},
+                      "match": {"location": "spec/cucumber/formatter/json_spec.rb:575"},
                       "result": {"status": "passed",
                                  "duration": 1}}]}]}]))
           end
@@ -661,31 +643,31 @@ module Cucumber
                    "description": "",
                    "type": "scenario",
                    "before":
-                    [{"match": {"location": "spec/cucumber/formatter/json_spec.rb:638"},
+                    [{"match": {"location": "spec/cucumber/formatter/json_spec.rb:620"},
                       "result": {"status": "passed",
                                  "duration": 1}},
-                     {"match": {"location": "spec/cucumber/formatter/json_spec.rb:639"},
+                     {"match": {"location": "spec/cucumber/formatter/json_spec.rb:621"},
                       "result": {"status": "passed",
                                  "duration": 1}}],
                    "steps":
                     [{"keyword": "Given ",
                       "name": "there are bananas",
                       "line": 4,
-                      "match": {"location": "spec/cucumber/formatter/json_spec.rb:645"},
+                      "match": {"location": "spec/cucumber/formatter/json_spec.rb:627"},
                       "result": {"status": "passed",
                                  "duration": 1},
                       "after":
-                       [{"match": {"location": "spec/cucumber/formatter/json_spec.rb:642"},
+                       [{"match": {"location": "spec/cucumber/formatter/json_spec.rb:624"},
                          "result": {"status": "passed",
                                     "duration": 1}},
-                        {"match": {"location": "spec/cucumber/formatter/json_spec.rb:643"},
+                        {"match": {"location": "spec/cucumber/formatter/json_spec.rb:625"},
                          "result": {"status": "passed",
                                     "duration": 1}}]}],
                    "after":
-                    [{"match": {"location": "spec/cucumber/formatter/json_spec.rb:641"},
+                    [{"match": {"location": "spec/cucumber/formatter/json_spec.rb:623"},
                       "result": {"status": "passed",
                                  "duration": 1}},
-                     {"match": {"location": "spec/cucumber/formatter/json_spec.rb:640"},
+                     {"match": {"location": "spec/cucumber/formatter/json_spec.rb:622"},
                       "result": {"status": "passed",
                                  "duration": 1}}]}]}]))
           end
@@ -726,13 +708,13 @@ module Cucumber
                     [{"keyword": "Given ",
                       "name": "there are bananas",
                       "line": 4,
-                      "match": {"location": "spec/cucumber/formatter/json_spec.rb:707"},
+                      "match": {"location": "spec/cucumber/formatter/json_spec.rb:689"},
                       "result": {"status": "passed",
                                  "duration": 1}}],
                    "around":
                     [{"match": {"location": "unknown_hook_location:1"},
                       "result": {"status": "failed",
-                                 "error_message": "error (RuntimeError)\\n./spec/cucumber/formatter/json_spec.rb:705:in `Around'",
+                                 "error_message": "error (RuntimeError)\\n./spec/cucumber/formatter/json_spec.rb:687:in `Around'",
                                  "duration": 1}}]}]}]})
           end
         end
@@ -773,7 +755,7 @@ module Cucumber
                        "rows":
                          [{"cells": ["aa", "bb"]},
                           {"cells": ["cc", "dd"]}],
-                       "match": {"location": "spec/cucumber/formatter/json_spec.rb:751"},
+                       "match": {"location": "spec/cucumber/formatter/json_spec.rb:733"},
                        "result": {"status": "passed",
                                   "duration": 1}}]}]}]))
           end